debt - remove goToDeclaration/common
